Add shared commands and move  to shared
