+++ Plugins/Wox.Plugin.WebSearch/Wox.Plugin.WebSearch.csproj
@@ -58,7 +58,6 @@
     <Compile Include="..\..\SolutionAssemblyInfo.cs">
       <Link>Properties\SolutionAssemblyInfo.cs</Link>
     </Compile>
-    <Compile Include="Commands\SearchWeb.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SearchSourceViewModel.cs" />
     <Compile Include="SettingsViewModel.cs" />
